requirements.txt
python-jobspy>=1.1.75
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
weasyprint>=60.0
jinja2>=3.1.0
pyyaml>=6.0
tenacity>=8.0.0
